simplifying changes
