+++ sql/connect/server/src/test/scala/org/apache/spark/sql/connect/service/SparkConnectBatchExecuteSuite.scala
@@ -14,349 +14,280 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.spark.sql.connect.service
 
 import java.util.UUID
 
 import scala.jdk.CollectionConverters._
 
-import org.scalatest.concurrent.Eventually
-
-import org.apache.spark.sql.AnalysisException
+import org.apache.spark.connect.proto
 import org.apache.spark.sql.connect.SparkConnectServerTest
 
 class SparkConnectBatchExecuteSuite extends SparkConnectServerTest {
 
-  test("Batch execute with multiple successful operations") {
+  private def createPlan(query: String): proto.Plan = {
+    proto.Plan
+      .newBuilder()
+      .setRoot(
+        proto.Relation
+          .newBuilder()
+          .setSql(proto.SQL
+            .newBuilder()
+            .setQuery(query)))
+      .build()
+  }
+
+  test("batch execute with multiple sequences") {
     withClient { client =>
-      val plan1 = buildPlan("SELECT 1 as value")
-      val plan2 = buildPlan("SELECT 2 as value")
-      val plan3 = buildPlan("SELECT 3 as value")
+      val seq1 = Seq((createPlan("SELECT 1"), None), (createPlan("SELECT 2"), None))
+      val seq2 = Seq((createPlan("SELECT 3"), None), (createPlan("SELECT 4"), None))
 
-      val response = client.batchExecute(
-        Seq((plan1, None), (plan2, None), (plan3, None)),
-        rollbackOnFailure = true)
+      val response = client.batchExecute(Seq(seq1, seq2))
 
-      assert(response.getResultsCount == 3)
-      assert(response.getResultsList.asScala.forall(_.getSuccess))
+      assert(response.getSequenceResultsCount == 2)
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(0).getQueryOperationIdsCount == 2)
+      assert(response.getSequenceResults(1).getSuccess)
+      assert(response.getSequenceResults(1).getQueryOperationIdsCount == 2)
 
-      // Verify operation IDs are unique and valid UUIDs
-      val operationIds = response.getResultsList.asScala.map(_.getOperationId)
-      assert(operationIds.size == 3)
-      assert(operationIds.distinct.size == 3)
-      operationIds.foreach { opId =>
-        UUID.fromString(opId) // Should not throw
+      // Verify we can reattach to each sequence
+      val seqOpId1 = response.getSequenceResults(0).getSequenceOperationId
+      val seqOpId2 = response.getSequenceResults(1).getSequenceOperationId
+
+      Thread.sleep(100) // Give sequences time to initialize
+
+      // Verify we can reattach to each sequence and they complete
+      val reattachIter1 = client.reattach(seqOpId1)
+      var hasQueryOpId1 = false
+      var hasResultComplete1 = false
+      while (reattachIter1.hasNext) {
+        val resp = reattachIter1.next()
+        if (resp.hasQueryOperationId) hasQueryOpId1 = true
+        if (resp.hasResultComplete) hasResultComplete1 = true
       }
+      assert(hasQueryOpId1)
+      assert(hasResultComplete1)
 
-      // Use reattach to verify operations complete successfully
-      // Give operations a moment to initialize before reattaching
-      Thread.sleep(100)
-      operationIds.foreach { opId =>
-        val reattachIterator = client.reattach(opId)
-        var hasResultComplete = false
-        while (reattachIterator.hasNext) {
-          val resp = reattachIterator.next()
-          if (resp.hasResultComplete) {
-            hasResultComplete = true
-          }
-        }
-        assert(hasResultComplete, s"Operation $opId did not complete successfully")
+      val reattachIter2 = client.reattach(seqOpId2)
+      var hasQueryOpId2 = false
+      var hasResultComplete2 = false
+      while (reattachIter2.hasNext) {
+        val resp = reattachIter2.next()
+        if (resp.hasQueryOperationId) hasQueryOpId2 = true
+        if (resp.hasResultComplete) hasResultComplete2 = true
       }
+      assert(hasQueryOpId2)
+      assert(hasResultComplete2)
     }
   }
 
-  test("Batch execute with client-provided operation IDs") {
+  test("batch execute with custom sequence operation IDs") {
     withClient { client =>
-      val opId1 = UUID.randomUUID().toString
-      val opId2 = UUID.randomUUID().toString
+      val seqOpId1 = UUID.randomUUID().toString
+      val seqOpId2 = UUID.randomUUID().toString
 
-      val plan1 = buildPlan("SELECT 1 as value")
-      val plan2 = buildPlan("SELECT 2 as value")
+      val seq1 = Seq((createPlan("SELECT 1"), None))
+      val seq2 = Seq((createPlan("SELECT 2"), None))
 
-      val response = client.batchExecute(
-        Seq((plan1, Some(opId1)), (plan2, Some(opId2))),
-        rollbackOnFailure = false)
+      val response = client.batchExecute(Seq(seq1, seq2), Seq(Some(seqOpId1), Some(seqOpId2)))
 
-      assert(response.getResultsCount == 2)
-      assert(response.getResults(0).getOperationId == opId1)
-      assert(response.getResults(1).getOperationId == opId2)
-      assert(response.getResults(0).getSuccess)
-      assert(response.getResults(1).getSuccess)
-
-      // Use reattach to verify operations complete successfully
-      // Give operations a moment to initialize before reattaching
-      Thread.sleep(100)
-      Seq(opId1, opId2).foreach { opId =>
-        val reattachIterator = client.reattach(opId)
-        var hasResultComplete = false
-        while (reattachIterator.hasNext) {
-          val resp = reattachIterator.next()
-          if (resp.hasResultComplete) {
-            hasResultComplete = true
-          }
-        }
-        assert(hasResultComplete, s"Operation $opId did not complete successfully")
-      }
+      assert(response.getSequenceResultsCount == 2)
+      assert(response.getSequenceResults(0).getSequenceOperationId == seqOpId1)
+      assert(response.getSequenceResults(1).getSequenceOperationId == seqOpId2)
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(1).getSuccess)
     }
   }
 
-  test("Batch execute with duplicate operation ID should fail") {
+  test("batch execute with custom query operation IDs") {
     withClient { client =>
-      val opId = UUID.randomUUID().toString
-      val plan1 = buildPlan("SELECT 1 as value")
-      val plan2 = buildPlan("SELECT 2 as value")
+      val queryOpId1 = UUID.randomUUID().toString
+      val queryOpId2 = UUID.randomUUID().toString
 
-      val response = client.batchExecute(
-        Seq((plan1, Some(opId)), (plan2, Some(opId))), // Duplicate ID
-        rollbackOnFailure = false)
+      val seq1 = Seq(
+        (createPlan("SELECT 1"), Some(queryOpId1)),
+        (createPlan("SELECT 2"), Some(queryOpId2)))
 
-      assert(response.getResultsCount == 2)
-      assert(response.getResults(0).getSuccess)
-      assert(!response.getResults(1).getSuccess)
-      assert(response.getResults(1).getErrorMessage.contains("OPERATION_ALREADY_EXISTS"))
+      val response = client.batchExecute(Seq(seq1))
+
+      assert(response.getSequenceResultsCount == 1)
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(0).getQueryOperationIdsCount == 2)
+
+      val queryOpIds = response.getSequenceResults(0).getQueryOperationIdsList.asScala
+      assert(queryOpIds.exists(_.getOperationId == queryOpId1))
+      assert(queryOpIds.exists(_.getOperationId == queryOpId2))
     }
   }
 
-  test("Batch execute with invalid plan and rollback enabled") {
+  test("batch execute with invalid sequence operation ID format") {
     withClient { client =>
-      val validPlan1 = buildPlan("SELECT 1 as value")
-      val invalidPlan = buildPlan("SELECT * FROM non_existent_table")
-      val validPlan2 = buildPlan("SELECT 2 as value")
+      val seq1 = Seq((createPlan("SELECT 1"), None))
 
-      val response = client.batchExecute(
-        Seq((validPlan1, None), (invalidPlan, None), (validPlan2, None)),
-        rollbackOnFailure = true)
-
-      // All should fail with rollback enabled
-      assert(response.getResultsCount == 3)
-
-      // The first and the last operation will complete successfully, but the second
-      // operation will fail. This means we reattach each operation and make sure it
-      // completes successfully except the second operation
-      // where the client will throw an exception.
-
-      // Reattach the first operation:
-      var reattachIterator = client.reattach(response.getResults(0).getOperationId)
-      var hasResultComplete = false
-      while (reattachIterator.hasNext) {
-        val resp = reattachIterator.next()
-        if (resp.hasResultComplete) {
-          hasResultComplete = true
-        }
-      }
-      assert(
-        hasResultComplete,
-        s"Operation ${response.getResults(0).getOperationId} did not complete")
-
-      // Reattach the second operation:
-      intercept[AnalysisException] {
-        reattachIterator = client.reattach(response.getResults(1).getOperationId)
-        while (reattachIterator.hasNext) {
-          reattachIterator.next()
-        }
-      }
-
-      // Reattach the third operation:
-      reattachIterator = client.reattach(response.getResults(2).getOperationId)
-      hasResultComplete = false
-      while (reattachIterator.hasNext) {
-        val resp = reattachIterator.next()
-        if (resp.hasResultComplete) {
-          hasResultComplete = true
-        }
-      }
-      assert(
-        hasResultComplete,
-        s"Operation ${response.getResults(2).getOperationId} did not complete")
-    }
-  }
-
-  test("Batch execute with invalid plan and rollback disabled") {
-    withClient { client =>
-      val validPlan1 = buildPlan("SELECT 1 as value")
-      val invalidPlan = buildPlan("SELECT * FROM non_existent_table")
-      val validPlan2 = buildPlan("SELECT 2 as value")
-
-      val response = client.batchExecute(
-        Seq((validPlan1, None), (invalidPlan, None), (validPlan2, None)),
-        rollbackOnFailure = false)
-
-      // First and third should succeed, second should fail
-      assert(response.getResultsCount == 3)
-
-      // The first and the last operation will complete successfully, but the second
-      // operation will fail. This means we reattach each operation and make sure it
-      // completes successfully except the second operation
-      // where the client will throw an exception.
-
-      // Reattach the first operation:
-      var reattachIterator = client.reattach(response.getResults(0).getOperationId)
-      var hasResultComplete = false
-      while (reattachIterator.hasNext) {
-        val resp = reattachIterator.next()
-        if (resp.hasResultComplete) {
-          hasResultComplete = true
-        }
-      }
-      assert(
-        hasResultComplete,
-        s"Operation ${response.getResults(0).getOperationId} did not complete")
-
-      // Reattach the second operation:
-      intercept[AnalysisException] {
-        reattachIterator = client.reattach(response.getResults(1).getOperationId)
-        while (reattachIterator.hasNext) {
-          reattachIterator.next()
-        }
-      }
-
-      // Reattach the third operation:
-      reattachIterator = client.reattach(response.getResults(2).getOperationId)
-      hasResultComplete = false
-      while (reattachIterator.hasNext) {
-        val resp = reattachIterator.next()
-        if (resp.hasResultComplete) {
-          hasResultComplete = true
-        }
-      }
-      assert(
-        hasResultComplete,
-        s"Operation ${response.getResults(2).getOperationId} did not complete")
-    }
-  }
-
-  test("Batch execute with empty list") {
-    withClient { client =>
-      val response = client.batchExecute(Seq.empty, rollbackOnFailure = false)
-
-      assert(response.getResultsCount == 0)
-      assert(response.getSessionId.nonEmpty)
-      assert(response.getServerSideSessionId.nonEmpty)
-    }
-  }
-
-  test("Batch execute preserves session context") {
-    withClient { client =>
-      // Create a temp view in the session
-      val createViewPlan = buildSqlCommandPlan("CREATE TEMP VIEW test_view AS SELECT 1 as id")
-      val createViewResponse = client.execute(createViewPlan)
-      while (createViewResponse.hasNext) createViewResponse.next()
-
-      // Use the temp view in batch execute
-      val plan1 = buildPlan("SELECT * FROM test_view")
-      val plan2 = buildPlan("SELECT id * 2 as double_id FROM test_view")
-
-      val response =
-        client.batchExecute(Seq((plan1, None), (plan2, None)), rollbackOnFailure = false)
-
-      assert(response.getResultsCount == 2)
-      assert(response.getResults(0).getSuccess)
-      assert(response.getResults(1).getSuccess)
-
-      // Use reattach to verify operations complete successfully and can access session view
-      // Give operations a moment to initialize before reattaching
-      Thread.sleep(100)
-      response.getResultsList.asScala.foreach { result =>
-        val reattachIterator = client.reattach(result.getOperationId)
-        var hasResultComplete = false
-        while (reattachIterator.hasNext) {
-          val resp = reattachIterator.next()
-          if (resp.hasResultComplete) {
-            hasResultComplete = true
-          }
-        }
-        assert(hasResultComplete, s"Operation ${result.getOperationId} did not complete")
-      }
-    }
-  }
-
-  test("Batch execute with command plans") {
-    withClient { client =>
-      val createView1 = buildSqlCommandPlan("CREATE TEMP VIEW v1 AS SELECT 1 as a")
-      val createView2 = buildSqlCommandPlan("CREATE TEMP VIEW v2 AS SELECT 2 as b")
-
-      val response = client.batchExecute(
-        Seq((createView1, None), (createView2, None)),
-        rollbackOnFailure = false)
-
-      assert(response.getResultsCount == 2)
-      assert(response.getResults(0).getSuccess)
-      assert(response.getResults(1).getSuccess)
-
-      // Use reattach to verify operations complete successfully
-      // Give operations a moment to initialize before reattaching
-      Thread.sleep(100)
-      response.getResultsList.asScala.foreach { result =>
-        val reattachIterator = client.reattach(result.getOperationId)
-        var hasResultComplete = false
-        while (reattachIterator.hasNext) {
-          val resp = reattachIterator.next()
-          if (resp.hasResultComplete) {
-            hasResultComplete = true
-          }
-        }
-        assert(hasResultComplete, s"Operation ${result.getOperationId} did not complete")
-      }
-
-      // Verify views were created
-      val queryPlan = buildPlan("SELECT * FROM v1 UNION ALL SELECT * FROM v2")
-      val queryResponse = client.execute(queryPlan)
-      var rowCount = 0
-      while (queryResponse.hasNext) {
-        queryResponse.next()
-        rowCount += 1
-      }
-      // Should have at least processed the query successfully
-    }
-  }
-
-  test("Batch execute operations run independently") {
-    withClient { client =>
-      // Create long-running queries
-      val plan1 = buildPlan("SELECT * FROM range(10000)")
-      val plan2 = buildPlan("SELECT * FROM range(10000)")
-      val plan3 = buildPlan("SELECT * FROM range(10000)")
-
-      val response = client.batchExecute(
-        Seq((plan1, None), (plan2, None), (plan3, None)),
-        rollbackOnFailure = false)
-
-      assert(response.getResultsCount == 3)
-      val operationIds = response.getResultsList.asScala.map(_.getOperationId)
-
-      // All operations should be successfully started
-      assert(response.getResultsList.asScala.forall(_.getSuccess))
-
-      // Verify operations are in the execution manager
-      Eventually.eventually(timeout(eventuallyTimeout)) {
-        val holders = SparkConnectService.executionManager.listExecuteHolders
-        // At least some should still be running or all may have completed
-        assert(holders.length >= 0) // Operations are independent
-      }
-    }
-  }
-
-  test("Batch execute with invalid UUID format for operation ID") {
-    withClient { client =>
-      val plan = buildPlan("SELECT 1")
-
-      // The client will throw an exception before the request is sent to the server
+      // Client-side validation should throw IllegalArgumentException
       val ex = intercept[IllegalArgumentException] {
-        client.batchExecute(Seq((plan, Some("not-a-valid-uuid"))), rollbackOnFailure = false)
+        client.batchExecute(Seq(seq1), Seq(Some("invalid-uuid")))
       }
-      assert(ex.getMessage.contains("Invalid operationId"))
+      assert(ex.getMessage.contains("Invalid sequence operation ID"))
     }
   }
 
-  test("Batch execute respects session ID") {
-    val sessionId = UUID.randomUUID().toString
-    withClient(sessionId = sessionId) { client =>
-      val plan = buildPlan("SELECT 1")
+  test("batch execute with duplicate sequence operation ID") {
+    withClient { client =>
+      val seqOpId = UUID.randomUUID().toString
 
-      val response = client.batchExecute(Seq((plan, None)), rollbackOnFailure = false)
+      val seq1 = Seq((createPlan("SELECT 1"), None))
+      val seq2 = Seq((createPlan("SELECT 2"), None))
 
-      assert(response.getSessionId == sessionId)
+      // First submission should succeed
+      val response1 = client.batchExecute(Seq(seq1), Seq(Some(seqOpId)))
+      assert(response1.getSequenceResults(0).getSuccess)
+
+      // Second submission with same ID should fail
+      val response2 = client.batchExecute(Seq(seq2), Seq(Some(seqOpId)))
+      assert(!response2.getSequenceResults(0).getSuccess)
+      assert(response2.getSequenceResults(0).getErrorMessage.contains("OPERATION_ALREADY_EXISTS"))
+    }
+  }
+
+  test("batch execute with failing query stops sequence") {
+    withClient { client =>
+      val seq1 = Seq(
+        (createPlan("SELECT 1"), None),
+        (createPlan("SELECT * FROM non_existent_table"), None),
+        (createPlan("SELECT 2"), None)
+      ) // This should not execute
+
+      val response = client.batchExecute(Seq(seq1))
+
+      assert(response.getSequenceResultsCount == 1)
+      assert(response.getSequenceResults(0).getSuccess) // Submission succeeds
+      assert(response.getSequenceResults(0).getQueryOperationIdsCount == 3)
+
+      val seqOpId = response.getSequenceResults(0).getSequenceOperationId
+
+      Thread.sleep(100) // Give sequence time to start
+
+      // Reattach to see the error
+      val exception = intercept[Exception] {
+        val reattachIter = client.reattach(seqOpId)
+        while (reattachIter.hasNext) {
+          reattachIter.next()
+        }
+      }
+
+      assert(
+        exception.getMessage.contains("TABLE_OR_VIEW_NOT_FOUND") ||
+          exception.getMessage.contains("non_existent_table"))
+    }
+  }
+
+  test("batch execute with empty sequence list") {
+    withClient { client =>
+      val response = client.batchExecute(Seq.empty)
+
+      assert(response.getSequenceResultsCount == 0)
+    }
+  }
+
+  test("batch execute with single query in sequence") {
+    withClient { client =>
+      val seq1 = Seq((createPlan("SELECT 42"), None))
+
+      val response = client.batchExecute(Seq(seq1))
+
+      assert(response.getSequenceResultsCount == 1)
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(0).getQueryOperationIdsCount == 1)
+
+      val seqOpId = response.getSequenceResults(0).getSequenceOperationId
+
+      Thread.sleep(100)
+
+      val reattachIter = client.reattach(seqOpId)
+      var hasQueryOpId = false
+      var hasArrowBatch = false
+      while (reattachIter.hasNext) {
+        val resp = reattachIter.next()
+        if (resp.hasQueryOperationId) hasQueryOpId = true
+        if (resp.hasArrowBatch) hasArrowBatch = true
+      }
+      assert(hasQueryOpId)
+      assert(hasArrowBatch)
+    }
+  }
+
+  test("batch execute respects session ID") {
+    withClient { client =>
+      val seq1 = Seq((createPlan("SELECT 1"), None))
+
+      val response = client.batchExecute(Seq(seq1))
+
+      assert(response.getSessionId == client.sessionId)
       assert(response.getServerSideSessionId.nonEmpty)
+    }
+  }
+
+  test("batch execute with multiple queries in single sequence") {
+    withClient { client =>
+      val queries = (1 to 5).map(i => (createPlan(s"SELECT $i"), None))
+
+      val response = client.batchExecute(Seq(queries))
+
+      assert(response.getSequenceResultsCount == 1)
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(0).getQueryOperationIdsCount == 5)
+
+      // Verify query indices are correct
+      val queryOpIds = response.getSequenceResults(0).getQueryOperationIdsList.asScala
+      queryOpIds.zipWithIndex.foreach { case (qid, idx) =>
+        assert(qid.getQueryIndex == idx)
+      }
+    }
+  }
+
+  test("batch execute with mixed success and failure sequences") {
+    withClient { client =>
+      val seq1 = Seq((createPlan("SELECT 1"), None)) // Success
+      val seq2 = Seq((createPlan("INVALID SQL"), None)) // Will fail during execution
+      val seq3 = Seq((createPlan("SELECT 3"), None)) // Success
+
+      val response = client.batchExecute(Seq(seq1, seq2, seq3))
+
+      assert(response.getSequenceResultsCount == 3)
+      // All should submit successfully
+      assert(response.getSequenceResults(0).getSuccess)
+      assert(response.getSequenceResults(1).getSuccess)
+      assert(response.getSequenceResults(2).getSuccess)
+
+      Thread.sleep(100)
+
+      // seq1 should complete successfully
+      val reattachIter1 = client.reattach(response.getSequenceResults(0).getSequenceOperationId)
+      var hasResponses1 = false
+      while (reattachIter1.hasNext) {
+        reattachIter1.next()
+        hasResponses1 = true
+      }
+      assert(hasResponses1)
+
+      // seq2 should fail during execution (reattach will throw)
+      intercept[Exception] {
+        val reattachIter2 = client.reattach(response.getSequenceResults(1).getSequenceOperationId)
+        while (reattachIter2.hasNext) {
+          reattachIter2.next()
+        }
+      }
+
+      // seq3 should complete successfully
+      val reattachIter3 = client.reattach(response.getSequenceResults(2).getSequenceOperationId)
+      var hasResponses3 = false
+      while (reattachIter3.hasNext) {
+        reattachIter3.next()
+        hasResponses3 = true
+      }
+      assert(hasResponses3)
     }
   }
 }
